silently convert unicode to str for header values
header values must be `str` or `bytes`. But most of the strings in the
codebase are unicode, and it is bothering to have to explicetly convert
them to `str` for headers, I'd rather silently convert `unicode` value
to `str` instead of raising an exception.

The downside is that it's very python 2.x specific and will need to be
changed when making bambou python 3.x compliant.